neovoxel_bot/requirements.txt
python-dotenv
slackclient